Move common compose ui into their own module
This will allow such layout code or modifiers to be shared from normal
feature modules, but also from the design system module itself, which
won't have to depend on core-ui and create a circular dependency chain
